@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
 # A job to submit reports and run in background
-class ReportJob < ActiveJob::Base
+class ReportJob < ApplicationJob
   queue_as :default
 
   # The method used to run the report by rails.
   def perform(*reports)
     reports.each do |report|
       run_report(report)
-    rescue => e # rubocop:disable Style/RescueStandardError
-      report.update_attributes status: 'error' # rubocop:disable Rails/ActiveRecordAliases
+    rescue StandardError => e
+      report.update status: 'error'
       raise e
     end
   end
@@ -32,8 +32,8 @@ class ReportJob < ActiveJob::Base
                                                                 created_at: report.created_at })
       report.update! status: 'completed', output: output
     else
-      report.update_attributes status: 'error' # rubocop:disable Rails/ActiveRecordAliases
-      report.update_attributes status_message: r.error_message # rubocop:disable Rails/ActiveRecordAliases
+      report.update status: 'error'
+      report.update status_message: r.error_message
     end
   end
   # rubocop:enable Metrics/AbcSize, Metrics/MethodLength
