@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 # A basic generic report to be run
-class Report < ActiveRecord::Base
-  belongs_to :user
+class Report < ApplicationRecord
+  belongs_to :user, optional: true
   alias_attribute :creator, :user
   serialize :parameters
 
