@@ -3,7 +3,11 @@
 module ApplicationHelper
   # helper to toggle asc and desc ( for ordering )
   def switch_direction(direction)
-    direction.match?(/asc$/) ? [direction.gsub(/asc$/, 'desc'), 'arrow-up'] : [direction.gsub(/desc$/, 'asc'), 'arrow-down'] # rubocop:disable Metrics/LineLength
+    if direction.match?(/asc$/)
+      [direction.gsub(/asc$/, 'desc'), 'arrow-up']
+    else
+      [direction.gsub(/desc$/, 'asc'), 'arrow-down']
+    end
   end
 
   SORT_MAP = { hourly_rate: :hourly_rate_cents,
@@ -23,19 +27,27 @@ module ApplicationHelper
     SORT_MAP.with_indifferent_access[column] || column
   end
 
-  def multi_sort_link(column, title, direction = 'arrow-up') # rubocop:disable Metrics/AbcSize
-    # first we extract any existing sorts in the params
-    attrs = Array.wrap(params[:sort]).compact
+  def permitted_params
+    params.permit(:page, sort: [])
+  end
 
+  def sort_params(column, direction)
+    attrs = Array.wrap(permitted_params[:sort]).compact
     # now scan and look to see if we're already sortin on this column
     index = attrs.index { |a| a.match(/^#{Regexp.escape(column)} /) }
     if index
       attrs[index], direction = switch_direction(attrs[index])
     else
       # so we just stick it in the end
-      attrs << "#{column} asc"
+      attrs << "#{column} desc"
     end
-    link_to(title, params.merge(sort: attrs.compact), class: direction)
+    [attrs, direction]
+  end
+
+  def multi_sort_link(column, title, direction = 'arrow-up')
+    # first we extract any existing sorts in the params
+    attrs, direction = sort_params(column, direction)
+    link_to(title, permitted_params.merge(sort: attrs.compact), class: direction)
   end
 
   # Returns confirmation prompt text for delete action on the given object.
