@@ -11,10 +11,10 @@ module CasHelper
 
   def ensure_auth
     if session[:cas].nil? || session[:cas][:user].nil?
-      render status: :unauthorized, text: 'Redirecting to SSO...'
+      render status: :unauthorized, plain: 'Redirecting to SSO...'
     else
       update_current_user(User.find_by(cas_directory_id: session[:cas][:user]))
-      render status: :forbidden, text: 'Unrecognized user' unless @current_user
+      render status: :forbidden, plain: 'Unrecognized user' unless @current_user
     end
     nil
   end
