@@ -16,7 +16,8 @@ class Organization < ApplicationRecord
     end
   end
 
-  belongs_to :parent, class_name: 'Organization', foreign_key: 'organization_id', inverse_of: :children
+  belongs_to :parent, class_name: 'Organization', foreign_key: 'organization_id',
+                      optional: true, inverse_of: :children
   validates :organization_id, presence: true, unless: :root?
 
   has_many :children, foreign_key: :organization_id, class_name: 'Organization', inverse_of: :parent,
@@ -24,7 +25,11 @@ class Organization < ApplicationRecord
   # a more basic AR way of getting children
   has_many :organizations, dependent: :restrict_with_error
 
-  # rubocop:disable Rails/HasManyOrHasOneDependent - not sure what the right value is here
+  # a weird one. Rubocop wants use to have a dependent, but we don't want one
+  # actually. This is bc our has_one here is just some "association trickery",
+  # because the forign_key is being set to the cutoff :organization_type. When
+  # the org record is deleted, we shouldn't do anything to the Cutoff.
+  # rubocop:disable Rails/HasManyOrHasOneDependent
   has_one :organization_cutoff, foreign_key: :organization_type, primary_key: :organization_type,
                                 inverse_of: :organizations
   # rubocop:enable Rails/HasManyOrHasOneDependent
