@@ -1,2 +1,2 @@
-APP_VERSION = '1.2.0-115-g7beaced'
-GIT_TAG = '7beaced7e885cf2a656f13e818ad29253fdf10e5'
+APP_VERSION = '2.0.1-60-ga79ef94'
+GIT_TAG = 'a79ef946f30183817f2f61724f9f620389c9c24c'
