@@ -9,7 +9,7 @@ SimpleCov.formatters = [
 SimpleCov.start 'rails'
 
 ENV['RAILS_ENV'] ||= 'test'
-require File.expand_path('../config/environment', __dir__)
+require_relative '../config/environment'
 require 'rails/test_help'
 require 'minitest/reporters'
 Minitest::Reporters.use!
